+++ ChangeLog.md
@@ -9,6 +9,7 @@ Changes since 2.1.6 go here.
 New Features:
 
 * Added additional sanity checks when calling `Table.to_hdf5`, see [PR #769](https://github.com/biocore/biom-format/pull/769).
+* `Table.subsample()` can optionally perform subsampling with replacement. See [issue #774](https://github.com/biocore/biom-format/issues/774).
 
 Bug fixes: